Added basic sanitisation of email contents
When viewing the contents of an email (via HTML), all fields are now
sanitized of special HTML characters to prevent glitches when using
these characters in an email field.

The sanitisation is then reveresed if the email is replied to, so that
the email contents displays correctly inside the HTML forms.