Added Search functionality to all mailbox views
There is now a search bar that takes input text for all mailboxes.
On keyup the current mailbox is searched for emails with a sender
or subject that match the search query string, and only these emails
are displayed.

When the search bar is emptied manually or by clicking the X next
to the search bar, the full email inbox is restored.